refactor: move everything into a more structured hierarchy
One interesting thing to note is that we're just reading files with relative
paths now - regardless of where they are read from.  This seems to rely on
magic logic with Deno.cwd() being the assumed root for any relative file that
is read with Deno.readFile() or Deno.readFileSync().

While this feels strange, I suspect it'll help in the long term when we want
to create a static binary to ship.